requirements.txt
go-template==0.0.3
Jinja2==3.1.2
MarkupSafe==2.1.3
PyYAML==6.0.1
